// src/types/user.ts
export type Profile = {
  address: string | null;
  avatar_url: string | null;
  first_name: string | null;
  id: string;
  last_name: string | null;
  latitude: number | null;
  location: unknown | null;
  longitude: number | null;
  phone: string | null;
  search_radius: number | null;
  unit: string | null;
  updated_at: string | null;
  saved_places: string | null;
};